backend/requirements.txt
fastapi
uvicorn
google-cloud-speech
google-generativeai
pydantic
python-multipart
aiofiles